requirements.txt
streamlit>=1.33.0,<2.0.0
yfinance>=0.2.40
openai>=1.40.0
python-dotenv>=1.0.1
pandas>=2.0.0
